M1: give notes a stable authored EventId
Note ids were a per-clip uint32 counter starting at 1, so the first note of
every clip in a project had id 1. Ids were not unique even within one
document, which makes them useless for naming anything — and naming things is
the precondition for overrides that survive regeneration, for edit-preserving
generation, and for saying which notes an agent wrote.

EventId is now u64: author in the top 16 bits, a per-author counter in the low
48. Collisions become impossible by construction rather than by coordination,
and provenance is a bit test rather than a side table. 48 bits is ~281
trillion events per author.

An id authored elsewhere no longer perturbs a clip's counter: reserveNoteId
ignores ids from other authors, so two authors editing one document keep
separate counter spaces instead of racing one shared high-water mark.

UiClipNote grows 24 -> 32 bytes and kShmVersion goes 7 -> 8. Both sides assert
the layout — a static_assert in C++, offset assertions in the Rust mirror — so
a mismatch fails at compile time on one end and in tests on the other.

Chord ids are deliberately left at u32. UiChordDiffPayload is already exactly
40 bytes, the full EventEntry payload budget, so widening chordId needs a wire
redesign rather than a field change. Notes are what the duration work needs.

Not yet done: the Rust UI's ClipNote does not carry the id at all, so
clip_store remains keyed by position. Widening the wire is the precondition;
carrying identity into the UI model is the next step.

Verified end to end — three notes written through the CLI reach project.json
as ids 1, 2, 3 with author 0. Along the way the version guard added in M0
earned itself: a stale CLI refused the new layout with a clear mismatch
message instead of misreading a 32-byte struct as 24.

+++ apps/project_file.cpp
@@ -622,7 +622,7 @@ bool deserializeProject(const std::string& json,
               static_cast<uint8_t>(noteTree.get<uint32_t>("velocity", 0));
           event.payload.note.column =
               static_cast<uint8_t>(noteTree.get<uint32_t>("column", 0));
-          event.payload.note.noteId = noteTree.get<uint32_t>("note_id", 0);
+          event.payload.note.noteId = noteTree.get<uint64_t>("note_id", 0);
           track.clip.addEvent(event);
         }
       }
